allows packing plan to be predicted on the basis of current traffic or predicted traffic
Previously, we could only propose a new packing plan for a job if its current traffic was too high. In this diff, we allow the packing planner to be supplied with a TrafficProvider object. This object can either provide current traffic (as in the previous implementation) or it can provide traffic predicted for a future time instance. Given future traffic arrival rates and current performance metrics, the packing planner proposes a new packing plan for the job.

This commit also removes some extraneous code.

Differential Revision: https://phabricator.twitter.biz/D198285